When the bundle info (requirements, exports, symbolic name, version etc.) are not available, the dependency tracker and the process of recoding that in files does not break now. Empty placeholders for those values are created in the files.
Also,

The infrastructure now supports the analysis of plugins and features
that are not archives (jar files) and can be folderds as such (e.g. in
case of features, the metadata may be unpacked and stored as folders and
also for plugin packages / plugins that have to be unpacked -> extracted
from jars to be saved as directories for say providing mutimedia
assets).