separated out constants to be used for building plugin extracts and then later for use in parsing during rebuilding information from extracts.
started with the dependency finding functionality: DependencyFinder -
still work in progress.

also brought the PluginAnalysisJobRunner to a new package inside the
application with  a dummy example outside in src/examples . These need
to be refined later on.   This is because at some point in time we would
need a Job Runner as a part of the application.