tried resolving the Manifest.mf parsing error. more robust err handling needed in ManifestParser still.
----
the error where a manifest.mf file may not be parsed by the
ManifestParser because of some malformedness of some property or when
some property is missing, at least for the bundleDescription property in
case it is missing.
The problem was in ManifestParser when trying to parse bundle
description, so now now   removed the capability to parse foe bundle
decsription. A more robust exception handling on ManifestParser is
needed that apparantly apache.ivy does not do.

Also fixed some bug s:

> in DependencyFinder.java  when building exporters list and the
considerBundleExportersOnly flag is set to true. The split on the
invokation to get the invokation without the returnType was wrongly done
before. corrected now.

> in BundleAnalyser.java, the Bundle Exports property was not marked
with a terminator when recording that to the Extract file for plugin.