for ManifestParser to not break if some Manifest Attr is missing from manifest
The ordeal of exception handling in ManifestParser to not break if some
Manifest Attr is missing from manifest. Now the ManifestParser, moves
over to parsing for the next available attribute instead of breaking if
say BUNDLE_SYMBOLIC_NAME or BUNDLE_IMPORT_PACKAGES is not available.

The manifest parser will still throw an exception if such a case is
encountered where the format of the attribute's value itself is not
proper. In this case we cannot accept anything like eclipse does from
the manifest.mf file as now we do not know really what is correct and
what is not.