Now capturing potential proxies for invokation satisfaction in BundleAnalyser
the potential proxies being written to plugin extracts. The proxies are
modified invokation signatures with the classname of the invokation
replaced with the  uppermost in  that class's inheritence tree available
inside the plugin or the interface(s) implemented, in case this
invokation was not satisfiable directly or indirectly through interface
and inheritence sequencesat the plugin level.

these can be used to check for satisfaction at the DependencyFinder
level when checking for indirect cross plugin class dependencies.