Changed the  indirect dependency satisfaction process for imports, in the DependencyFinder to account for multiple plugins pointing to the same super class for a given class of a method import signature and so clubbing the various plugin options into one (A || B || C ..) ....
The places where multiple versions of a certain plugin or even many
plugins leading to the same inheritance hierarchy are clubbed with OR
together as any of them are okay to satisfy that class requirement and
then the parent class is then checked only once and not every single
time for all these versions / plugins, reducing the explosion of
combinations.

The constraint formation is now more optimized, therefore.
  thereby significantly reducing the length of constraints